+++ src/components/dashboard/HabitGoalStatus.vue
@@ -5,29 +5,34 @@ import { isWithinInterval, startOfDay, endOfDay } from 'date-fns';
 import { Goal } from 'src/lib/api/goal.ts';
 import { Tally } from 'src/lib/api/tally.ts';
 
-import { analyzeStreaksForHabit, GoalHabitParameters } from 'server/lib/models/goal.ts';
+import { GOAL_CADENCE_UNIT_INFO, analyzeStreaksForHabit, GoalHabitParameters } from 'server/lib/models/goal.ts';
 import { streakColors } from 'src/lib/tally.ts';
 import { formatDate, parseDateStringSafe } from 'src/lib/date.ts';
+import { commaify, formatPercent } from 'src/lib/number.ts';
 
 
 import GoalCard from 'src/components/dashboard/GoalCard.vue';
 import HabitGauge from 'src/components/goal/HabitGauge.vue';
+import StatTile from 'src/components/goal/StatTile.vue';
 
 const props = defineProps<{
   goal: Goal,
   tallies: Tally[],
 }>();
 
+const parameters = computed(() => {
+  return props.goal.parameters as GoalHabitParameters;
+})
+
 const habitStats = computed(() => {
   const now = new Date();
   const goalEndDate = parseDateStringSafe(props.goal.endDate) ?? now;
   const endDate = now < goalEndDate ? now : goalEndDate;
 
-  const parameters = props.goal.parameters as GoalHabitParameters;
   const stats = analyzeStreaksForHabit(
     props.tallies,
-    parameters.cadence,
-    parameters.threshold,
+    parameters.value.cadence,
+    parameters.value.threshold,
     props.goal.startDate,
     formatDate(endDate),
   );
@@ -35,7 +40,18 @@ const habitStats = computed(() => {
   return stats;
 });
 
-const ranges = computed(() => habitStats.value.ranges.toReversed().slice(0, 5));
+const successfulRanges = computed(() => {
+  return habitStats.value.ranges.filter(range => range.isSuccess);
+});
+
+const currentRange = computed(() => {
+  const currentRanges = habitStats.value.ranges.filter(range => rangeContainsToday(range));
+  if(currentRanges.length > 0) {
+    return currentRanges[0];
+  } else {
+    return null;
+  }
+})
 
 function rangeContainsToday(range) {
   const now = new Date();
@@ -53,21 +69,52 @@ function rangeContainsToday(range) {
   >
     <template #comment>
       <div
-        v-if="habitStats.streaks.current.length > 1"
+        v-if="habitStats.streaks.current !== null && habitStats.streaks.current.length > 1"
         :class="[ 'px-2 py-1 rounded-full text-sm font-normal', streakColors(habitStats.streaks.current) ]"
       >
-        {{ habitStats.streaks.current }} in a row!
+        {{ habitStats.streaks.current.length }} in a row!
       </div>
     </template>
     <template #content>
-      <div class="flex gap-1 overflow-x-auto">
-        <HabitGauge
-          v-for="range of ranges"
-          :key="range.startDate"
-          :range="range"
-          :goal="props.goal"
-          :highlight="rangeContainsToday(range)"
-        />
+      <div class="habit-summary flex flex-wrap justify-evenly md:justify-start items-center gap-4 mb-4">
+        <div
+          v-if="currentRange !== null"
+          class="current-progress"
+        >
+          <HabitGauge
+            :goal="props.goal"
+            :range="currentRange"
+          />
+        </div>
+        <div
+          v-if="habitStats.streaks.current !== null"
+          class="current-streak"
+        >
+          <StatTile
+            top-legend="Current streak"
+            :highlight="commaify(habitStats.streaks.current.length)"
+            :suffix="GOAL_CADENCE_UNIT_INFO[parameters.cadence.unit].label[habitStats.streaks.current.length === 1 ? 'singular' : 'plural']"
+            bottom-legend="in a row"
+          />
+        </div>
+        <div
+          v-if="habitStats.streaks.longest !== null"
+          class="current-streak"
+        >
+          <StatTile
+            top-legend="Longest streak"
+            :highlight="commaify(habitStats.streaks.longest.length)"
+            :suffix="GOAL_CADENCE_UNIT_INFO[parameters.cadence.unit].label[habitStats.streaks.longest.length === 1 ? 'singular' : 'plural']"
+            bottom-legend="in a row"
+          />
+        </div>
+        <div class="success-rate">
+          <StatTile
+            :top-legend="`Success rate`"
+            :highlight="`${formatPercent(successfulRanges.length, habitStats.ranges.length)}%`"
+            :bottom-legend="`or ${successfulRanges.length} / ${habitStats.ranges.length}`"
+          />
+        </div>
       </div>
     </template>
   </GoalCard>