+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Types of changes include:
 
 ## Upcoming/Unreleased
 
+- FIXED: The streak counter for habit goals on the dashboard no longer dumps a bunch of unrelated text and instead now gives a cheery "X in a row!" counter like it's supposed to. (h/t CaitSidhe, sarah, Quinoafox)
+
 ## 0.15.0
 - NEW: Public profiles have launched! You can now enable your public profile in Settings. Once you do, your public profile will be available at trackbear.app/@[your username]. Profiles list lifetime stats by default, and you can add goals and projects to it by going to the goal or project, clicking **Configure**, and toggling the switch under "Show on profile?".